+++ src/main/java/domain/piece/Piece.java
@@ -2,6 +2,7 @@ package domain.piece;
 
 import domain.path.PathInfo;
 import domain.board.Position;
+import domain.piece.strategy.MoveStrategy;
 
 import java.util.List;
 
@@ -10,7 +11,7 @@ public record Piece(Camp camp, PieceType pieceType, MoveStrategy moveStrategy) {
         return new Piece(camp, pieceType, pieceType.moveStrategy());
     }
 
-    public boolean isSameCampe(Piece otherPiece){
+    public boolean isSameCampe(Piece otherPiece) {
         return otherPiece.camp.equals(camp);
     }
 
@@ -18,7 +19,7 @@ public record Piece(Camp camp, PieceType pieceType, MoveStrategy moveStrategy) {
         return moveStrategy.getPath(departure, destination);
     }
 
-    public void validateBlockingPiece(List<PathInfo> pathInfos, Position departure, Position destination){
+    public void validateBlockingPiece(List<PathInfo> pathInfos, Position departure, Position destination) {
         moveStrategy.validateBlockingPiece(pathInfos, departure, destination);
     }
 }
